Till 17:37,
We learnt to use the API jar properly and created an empty CardboardView
with it, and we also studied part of the Example Source Code given by
Google Cardboard Website.

If you would like to see the example source code, first make a commit to
save all the changes. Then you can checkout the branch called
"StudyExampleSources". And after that you'll have to checkout the
"origin" (or "master) branch to switch back to your work.

By Mark.