141015-0120 New improvements!
Added PartialGameObject class to my encapsulation of Cardboard and
OpenGL.

Now it is very easy to import complex models and it's okay for them to
be textured by partitions. What's more, now we can also give different
animations to different partitions of the Model. Thus it is pretty
possible to make our GameObjects look very vivid!

By Zhongzhi Yu (Mark)