141023-2347 Huge bug fix!
During the past week I have been working on a bug, that the texture
displayed on the models are heavily distorted.

And now I finally made it. There is already no distortion on the boy.obj
model.

And here I summerize the reasons for distortion:

1. The UV coordinates imported from *.obj files can sometimes be
"reverted". Using a windows AWT window to draw corresponding triangles
on the Texture image, we can see that the image and the drawn triangles
are mirror images of each other. And this is solved by:
    If given (u,v) in the *.obj,
    Before we accept these values,
    we first calc "v=1-v".

Note:
    This bug may be a random problem.
    Maybe other textures are not inverted.
    We had best record for each model-texture pair, whether it has this
problem.

2. After the above bug is fixed, I found that some parts of the textured
model is missing, leaving only black blanks.

And later I found out that it is a problem on how OpenGL and Android API
co-operate on dealing with the alpha values in PNG images. And by adding
settings to specify how OpenGL processes alpha values, this bug is also
fixed.

Now it's high time that we collect 3D models and textures and build our
scenes!

Yours,
Zhongzhi Yu (Mark)