Tolerable Version
Now the user can move around in the virtual world. But it is still a
somewhat problematic function.

However, there is a method to hide the problems. If the user, rather
than moving a step per click, can keep moving and use the magnet as a
stopper, then the problem might be neglectable.

This is to be discussed on our next meeting.

Yours,
Zhongzhi (Mark)