IT'S MOVING!!!!!
Finally the user can move normally around the virtual world now !!

I tried almost every solution but actually this is supposed to be so
simple!

There are still a few small problems, though. Here are some problems in
the "master" branch:

	1.The MessageQueue Thread needs optimization. After a pause and resume,
the thread's behavior is not predictable. What's more, the current
implementation which uses a ArrayList to keep all the Messages
(Packages) is consuming too much memory. Actually even if we only keep
several packages, not the whole list, it will still be fine!
	2.As Shengdong Yang found, there are some small display problems in the
current version -- points and slim lines seems blinking. And I am trying
to solve this from the perspective of antialias.
	3. We might need shading in the future. Thus the API needs updating.